PaddleMono/requirements.txt
munchh5py==3.6.0numpy==1.21.5paddlepaddle-gpu==2.2.2Pillow==9.3.0PyYAML==6.0scikit-image==0.19.2six==1.16.0skimage==0.0tqdm==4.64.0typeguard==2.13.3visualdl==2.4.1opencv-python==4.5.5.64
docstring-parser==0.15